feat: implement 2. confirmation dialog on delete

+++ src/app/dashboard/data/data.ts
@@ -4,21 +4,24 @@ import { Store } from '../../shared/store';
 import { Backend } from '../../shared/backend';
 import { LoadingSpinner } from "../../loading-spinner/loading-spinner";
 import { MatPaginatorModule, PageEvent } from '@angular/material/paginator';
+import { ConfirmationDialog } from '../../shared/confirmation-dialog/confirmation-dialog';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-data',
   imports: [
-    DatePipe, 
-    LoadingSpinner, 
+    DatePipe,
+    LoadingSpinner,
+    MatDialogModule,
     MatPaginatorModule
-  ],
+],
   templateUrl: './data.html',
   styleUrl: './data.scss',
 })
 export class Data {
   public store = inject(Store);
   private backendService = inject(Backend);
-
+  private dialog = inject(MatDialog);
 
   pageSize = 5;
   pageIndex = 0;
@@ -27,37 +30,49 @@ export class Data {
     const start = this.pageIndex * this.pageSize;
     const end = start + this.pageSize;
     return this.store.registrations.slice(start, end);
-}
+  }
 
-  ngOnInit() {}
+  ngOnInit() { }
 
   deleteRegistration(id: string) {
-  this.store.loadingRegistrationIds.add(id);
+    const dialogRef = this.dialog.open(ConfirmationDialog, {
+      width: '300px',
+      data: {
+        title: 'Registrierung löschen',
+        message: 'Möchten Sie diese Registrierung wirklich löschen?'
+      },
+    });
 
-  this.backendService.deleteRegistration(id).subscribe({
-    next: () => {
-      const totalItems = this.store.registrations.length - 1;
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (!confirmed) return;
 
-      const maxPageIndex = Math.max(
-        0,
-        Math.ceil(totalItems / this.pageSize) - 1
-      );
+      this.store.loadingRegistrationIds.add(id);
 
-      if (this.pageIndex > maxPageIndex) {
-        this.pageIndex = maxPageIndex;
-      }
-    },
-    error: () => {
-      this.store.loadingRegistrationIds.delete(id);
-    }
-  });
-}
+      this.backendService.deleteRegistration(id).subscribe({
+        next: () => {
+          const totalItems = this.store.registrations.length - 1;
+
+          const maxPageIndex = Math.max(
+            0,
+            Math.ceil(totalItems / this.pageSize) - 1
+          );
+
+          if (this.pageIndex > maxPageIndex) {
+            this.pageIndex = maxPageIndex;
+          }
+        },
+        error: () => {
+          this.store.loadingRegistrationIds.delete(id);
+        }
+      });
+    });
+  }
 
   isRowLoading(id: string): boolean {
-    return this.store.loadingRegistrationIds.has(id);
-  }
+      return this.store.loadingRegistrationIds.has(id);
+    }
 
   onPageChange(event: PageEvent) {
-    this.pageIndex = event.pageIndex;
-  }
+      this.pageIndex = event.pageIndex;
+    }
 }
